+++ modules/post/components/button/ConfirmCancelButton.stories.tsx
@@ -1,5 +1,5 @@
 import { View } from 'react-native';
-import { ComponentMeta, ComponentStory } from '@storybook/react-native';
+import { ComponentMeta } from '@storybook/react-native';
 
 import { FontLoader } from '@shared/components/ThemeProvider';
 import ConfirmCancelButton from './ConfirmCancelButton';
