+++ modules/post/screens/MeetingPostCreator.tsx
@@ -5,7 +5,7 @@ import MeetingNameInput from '@post/components/creation/meetingNameInput/Meeting
 import Uploader from '@post/components/creation/uploader/Uploader';
 import VotingTimeRangePicker from '@post/components/creation/votingTimeRangePicker/VotingTimeRangePicker';
 import ScreenLayout from '@shared/components/layout/Layout';
-import ConfirmCancelButton from '@post/components/card/button/ConfirmCancelButton';
+import ConfirmCancelButton from '@post/components/button/ConfirmCancelButton';
 import useStyles from './style';
 
 export default function MeetingPostCreator() {
